:truck: api: move find-user-certificate route and controller from lib to src
