+++ api/src/certification/results/application/certification-route.js
@@ -1,0 +1,21 @@
+import { certificationController } from './certification-controller.js';
+
+const register = async function (server) {
+  server.route([
+    {
+      method: 'GET',
+      path: '/api/certifications',
+      config: {
+        handler: certificationController.findUserCertifications,
+        notes: [
+          '- **Route nécessitant une authentification**\n' +
+            '- Récupération de toutes les certifications complétées de l’utilisateur courant',
+        ],
+        tags: ['api', 'certifications'],
+      },
+    },
+  ]);
+};
+
+const name = 'certification-api';
+export { name, register };
