+++ api/tests/integration/application/certifications/index_test.js
@@ -7,22 +7,11 @@ describe('Integration | Application | Route | Certifications', function () {
   let httpTestServer;
 
   beforeEach(async function () {
-    sinon.stub(certificationController, 'findUserCertifications').returns('ok');
     sinon.stub(certificationController, 'getCertification').callsFake((request, h) => h.response('ok').code(200));
     sinon.stub(securityPreHandlers, 'hasAtLeastOneAccessOf').returns(() => true);
 
     httpTestServer = new HttpTestServer();
     await httpTestServer.register(moduleUnderTest);
-  });
-
-  describe('GET /api/certifications', function () {
-    it('should exist', async function () {
-      // when
-      const response = await httpTestServer.request('GET', '/api/certifications');
-
-      // then
-      expect(response.statusCode).to.equal(200);
-    });
   });
 
   describe('GET /api/certifications/:id', function () {
